requirements.txt
requests
bs4
django==1.11.7
django-registration==2.3
django-mptt==0.8.7
Pillow==4.3.0
sorl-thumbnail==12.3
django-image-cropping==1.1.0
-e git+git@github.com:SmileyChris/easy-thumbnails.git#egg=easy-thumbnails
django-celery==3.2.1
python_slugify==1.2.4
psycopg2
django-widget-tweaks==1.4.1
django-jsonview==1.1.0
django-rest-framework==3.7.3